+++ src/Core/Dfe.Complete.Domain/Enums/NoteTaskIdentifier.cs
@@ -200,4 +200,8 @@ public enum NoteTaskIdentifier
     [Description("redact_and_send")]
     [DisplayDescription("Redact and send documents")]
     RedactAndSend = 49,
+
+    [Description("confirm_statutory_consultation")]
+    [DisplayDescription("Confirm statutory consultation is complete and any issues are being managed")]
+    ConfirmStatutoryConsultation = 50,
 }